+++ ts/test/scenario.test.ts
@@ -6,11 +6,11 @@ import {
   Signal,
   SignalCb,
 } from "@holochain/client";
-import { readFileSync } from "node:fs";
 import yaml from "js-yaml";
-import assert from "node:assert";
-import { test } from "node:test";
+import { readFileSync } from "node:fs";
+import { assert, test, expect } from "vitest";
 import {
+  AppWithOptions,
   CONDUCTOR_CONFIG,
   Scenario,
   dhtSync,
@@ -24,8 +24,10 @@ const TEST_ZOME_NAME = "coordinator";
 test("runScenario - Install hApp bundle and access cells through role ids", async () => {
   await runScenario(async (scenario: Scenario) => {
     const alice = await scenario.addPlayerWithApp({
-      type: "path",
-      value: FIXTURE_HAPP_URL.pathname,
+      appBundleSource: {
+        type: "path",
+        value: FIXTURE_HAPP_URL.pathname,
+      },
     });
     assert.ok(alice.namedCells.get("test"));
   });
@@ -34,37 +36,43 @@ test("runScenario - Install hApp bundle and access cells through role ids", asyn
 test("runScenario - Catch error when calling non-existent zome", async () => {
   await runScenario(async (scenario: Scenario) => {
     const alice = await scenario.addPlayerWithApp({
-      type: "path",
-      value: FIXTURE_HAPP_URL.pathname,
+      appBundleSource: {
+        type: "path",
+        value: FIXTURE_HAPP_URL.pathname,
+      },
     });
 
-    await assert.rejects(
+    await expect(
       alice.cells[0].callZome<EntryHash>({
         zome_name: "NOZOME",
         fn_name: "create",
       }),
-    );
+    ).rejects.toThrow();
   });
 });
 
 test("runScenario - Catch error when attaching a protected port", async () => {
   await runScenario(async (scenario: Scenario) => {
     const alice = await scenario.addPlayerWithApp({
-      type: "path",
-      value: FIXTURE_HAPP_URL.pathname,
+      appBundleSource: {
+        type: "path",
+        value: FIXTURE_HAPP_URL.pathname,
+      },
     });
 
-    await assert.rejects(
+    await expect(
       alice.conductor.attachAppInterface({ port: 300, allowed_origins: "*" }),
-    );
+    ).rejects.toThrow();
   });
 });
 
 test("runScenario - Catch error when calling a zome of an undefined cell", async () => {
   await runScenario(async (scenario: Scenario) => {
     const alice = await scenario.addPlayerWithApp({
-      type: "path",
-      value: FIXTURE_HAPP_URL.pathname,
+      appBundleSource: {
+        type: "path",
+        value: FIXTURE_HAPP_URL.pathname,
+      },
     });
 
     assert.throws(() =>
@@ -73,46 +81,51 @@ test("runScenario - Catch error when calling a zome of an undefined cell", async
   });
 });
 
-test("runScenario - Catch error that occurs in a signal handler", async () => {
-  await runScenario(async (scenario: Scenario) => {
-    let signalHandlerAlice: SignalCb | undefined;
-    const signalReceivedAlice = new Promise<Signal>((_, reject) => {
-      signalHandlerAlice = () => {
-        reject();
-      };
-    });
+test.sequential(
+  "runScenario - Catch error that occurs in a signal handler",
+  async () => {
+    await runScenario(async (scenario: Scenario) => {
+      let signalHandlerAlice: SignalCb | undefined;
+      const signalReceivedAlice = new Promise<Signal>((_, reject) => {
+        signalHandlerAlice = () => {
+          reject();
+        };
+      });
 
-    const alice = await scenario.addPlayerWithApp({
-      type: "path",
-      value: FIXTURE_HAPP_URL.pathname,
-    });
-    assert.ok(signalHandlerAlice);
-    assert.ok("on" in alice.appWs);
-    alice.appWs.on("signal", signalHandlerAlice);
+      const alice = await scenario.addPlayerWithApp({
+        appBundleSource: {
+          type: "path",
+          value: FIXTURE_HAPP_URL.pathname,
+        },
+      });
+      assert.ok(signalHandlerAlice);
+      assert.ok("on" in alice.appWs);
+      alice.appWs.on("signal", signalHandlerAlice);
 
-    const signalAlice = { value: "hello alice" };
-    alice.cells[0].callZome({
-      zome_name: TEST_ZOME_NAME,
-      fn_name: "signal_loopback",
-      payload: signalAlice,
-    });
+      const signalAlice = { value: "hello alice" };
+      alice.cells[0].callZome({
+        zome_name: TEST_ZOME_NAME,
+        fn_name: "signal_loopback",
+        payload: signalAlice,
+      });
 
-    await assert.rejects(signalReceivedAlice);
-  });
-});
+      await expect(signalReceivedAlice).rejects.toThrow();
+    });
+  },
+);
 
 test("Set custom network config", async () => {
   const scenario = new Scenario();
   const initiateIntervalMs = 10_000;
   const minInitiateIntervalMs = 20_000;
 
-  const alice = await scenario.addPlayerWithApp(
-    {
+  const alice = await scenario.addPlayerWithApp({
+    appBundleSource: {
       type: "path",
       value: FIXTURE_HAPP_URL.pathname,
     },
-    { networkConfig: { initiateIntervalMs, minInitiateIntervalMs } },
-  );
+    options: { networkConfig: { initiateIntervalMs, minInitiateIntervalMs } },
+  });
 
   const tmpDirPath = alice.conductor.getTmpDirectory();
   const conductorConfig = yaml.load(
@@ -144,8 +157,10 @@ test("Install hApp bundle and access cell by role name", async () => {
   const scenario = new Scenario();
 
   const alice = await scenario.addPlayerWithApp({
-    type: "path",
-    value: FIXTURE_HAPP_URL.pathname,
+    appBundleSource: {
+      type: "path",
+      value: FIXTURE_HAPP_URL.pathname,
+    },
   });
   assert.ok(alice.namedCells.get("test"));
   await scenario.cleanUp();
@@ -369,7 +384,7 @@ test("dhtSync - Fails if some Ops are not synced among all conductors", async ()
 
   // Bob never receives the entry
   try {
-    await dhtSync([alice, bob], alice.cells[0].cell_id[0]);
+    await dhtSync([alice, bob], alice.cells[0].cell_id[0], undefined, 5000);
     assert.fail();
   } catch {
     assert(true);
@@ -426,8 +441,10 @@ test("dhtSync - Fails if some Ops are not integrated in a conductor", async () =
 test("runScenario - call zome by role name", async () => {
   await runScenario(async (scenario: Scenario) => {
     const alice = await scenario.addPlayerWithApp({
-      type: "path",
-      value: FIXTURE_HAPP_URL.pathname,
+      appBundleSource: {
+        type: "path",
+        value: FIXTURE_HAPP_URL.pathname,
+      },
     });
 
     const result = (await alice.namedCells.get("test")?.callZome({
@@ -437,5 +454,37 @@ test("runScenario - call zome by role name", async () => {
     })) as ActionHash;
 
     assert.ok(result);
+  });
+});
+
+test("runScenario - add players and then install apps for them", async () => {
+  await runScenario(async (scenario) => {
+    const players = await scenario.addPlayers(2);
+    const appsWithOptions: AppWithOptions[] = [
+      { appBundleSource: { type: "path", value: FIXTURE_HAPP_URL.pathname } },
+      { appBundleSource: { type: "path", value: FIXTURE_HAPP_URL.pathname } },
+    ];
+    const [aliceApp, bobApp] = await scenario.installAppsForPlayers(
+      appsWithOptions,
+      players,
+    );
+    assert.deepEqual(aliceApp.agentPubKey, players[0].agentPubKey);
+    assert.deepEqual(bobApp.agentPubKey, players[1].agentPubKey);
+
+    const content = "test-content";
+    const createEntryHash = await aliceApp.cells[0].callZome<EntryHash>({
+      zome_name: TEST_ZOME_NAME,
+      fn_name: "create",
+      payload: content,
+    });
+
+    await dhtSync([aliceApp, bobApp], aliceApp.cells[0].cell_id[0]);
+
+    const readContent = await bobApp.cells[0].callZome<typeof content>({
+      zome_name: TEST_ZOME_NAME,
+      fn_name: "read",
+      payload: createEntryHash,
+    });
+    assert.equal(readContent, content);
   });
 });